requirements.txt
transformers
torch
PyYAML
jupyter
matplotlib
scikit-learn
ipykernel
